test(platform-core): add shipping module tests
